+++ .gitignore
@@ -3,3 +3,5 @@ nameMe API.code-workspace
 __pycache__/
 api_keys.db
 api_keys.db-journal
+app.log
+api.log
