+++ templates/api_gen.html
@@ -12,7 +12,6 @@
         </form>
         <p>Unique Key =<span style="font-weight:bolder;"> {{key}}</span><br></p>
     </td>
-    <div class="vl"></div>
     <td>
         <h1>Usages of this API:</h1>
         <p></p>
